python_server/requirements.txt
websockets==11.0.3
pyautogui==0.9.54
Pillow>=10.0.0
